Added the Excel File Reader
Why this tooks so long: Finding a way to read from the excel file was
actually quite easy. The time consuming part was figuring out why the
data was not coming in as expected. In a certian column there was text
that contained multiple lines. This was hard to find as Excel just
deletes the multiple lines and displays it as one if there isn't enoguh
room. Now that I know that this is the case I can begin working on a
solution to that.