Trying to fix issue where datetimes are not read from excel correctly.
I am trying to fix an issue where when I read a datetime in from excel I
only get a number. I found the cause now I am  trying to implement the
solution. This sheds light on the bigger problem I will be addressing
down the road in development that is, I only store data in the database
as text. After I get a minimum working version I will start looking to
fix that.