Removed the database in favor of reading from csv
Now that I am reading from csv I can git rid of the database and read in
memory from now on. This is huge as the program is incredibly fast now.
There is a know issue when reading from the csv that I get some of the
wrong row values.. I am actively looking in to that and may switch to
CSVTools.Excel