Fixed bug that caused input to be read in wrong
Removed CSVFile as I no longer need it since I am using CSVTools. Fixed a bug in CSVTools that caused the
input to be read in incorrectly. If a cell in the excel file conatined multiple Windows new line characters it would break. After fixing this issue the program now reads data exactly as it should. Everything is now displaying as it should.