Trying to write the new excel file
It may look like I have not accoplished anything but I have actually
been struggling to find a solution to writing a subset of rows to a csv
file. I have tried multiple solutions, although none of them are
sufficient enough. The main issue is escaping all the values so that
excel will read them correctly.