The ability to save the select data is now working
Finally figured out who to write to a csv file with accuracy. The
changes won't be able to be seen here on GitHub as I made the changes to
a local project called DataTables. By making the CsvWriter class public I
was able to access methods for writing the data from Rows.